refactor(player-portal): decompose ItemShopPicker into focused sibling files
ItemShopPicker.tsx shrinks from 971 → 255 lines. Domain logic extracted to:
- shop-utils.ts:       TypeFilter, ItemGroup, PriceState, parseItemName,
                       variantRank, filterMatchesByType, resolvePriceState,
                       extractPriceFromDocument, rarity style helpers
- useFitPageSize.ts:   ResizeObserver hook for dynamic grid page sizing
- ShopTile.tsx:        grid tile component
- ShopItemDetail.tsx:  full-document overlay (variant picker, buy button)

ItemShopPicker now delegates grid rendering to CompendiumPicker (inline,
no modal); ShopItemDetail remains absolute-positioned over the grid.
All existing tests pass unchanged.
